refactor(filestore,responsestore): serve both SQL backends from one store
Both pairs were mechanically identical apart from placeholders, the no-rows
sentinel and the Exec result shape. scanStoredFile and scanStoredResponseRow
lose their notFound parameter: with sqlx there is one sentinel to compare.

filestore's test carried its own four-backend harness keyed on TEST_DATABASE_DSN
and MONGO_TEST_DSN, where the postgres case had no CI path. It now shares one
behavioural suite across memory, both SQL dialects and Mongo, and covers the
not-found and validation paths it previously skipped.

responsestore's suite moves onto sqlxtest: 7 PostgreSQL subtests now run where
none did before.
